Adding fsdp fabric tests to pipeline
