+++ .azure/hpu-tests.yml
@@ -114,7 +114,7 @@ jobs:
           tests/test_pytorch/test_dynamic_shapes.py \
           tests/test_pytorch/test_datamodule.py \
           tests/test_pytorch/test_profiler.py \
-          tests/test_pytorch/test_precision.py \
+          tests/test_pytorch/test_precision.py
           --hpus 1 -m "not standalone_only" \
           --junitxml=hpu_test-torch-results.xml
       displayName: 'HPU General tests'
@@ -127,6 +127,8 @@ jobs:
       displayName: 'HPU torch compile tests'
 
     - bash: |
+        python -m pytest -sv tests/test_fabric/test_fsdp.py \
+          --hpus 1 --junitxml=hpu_test_fsdp-fabric-results.xml
         python -m pytest -sv tests/test_pytorch/test_fsdp.py \
           --hpus 1 --junitxml=hpu_test_fsdp-results.xml
       env:
@@ -162,7 +164,8 @@ jobs:
 
     - bash: |
         bash tests/run_standalone_tests.sh --hpus 2 -f \
-          tests/test_pytorch/test_fsdp.py
+          tests/test_pytorch/test_fsdp.py \
+          tests/test_fabric/test_fsdp.py
       env:
         PT_HPU_LAZY_MODE: 0
       displayName: 'FSDP Multi card(2) HPU test'
